#830 moved more things around
#1169
updated the log code to support getting the client address

+++ src/connection/sqlrconnection/bindformat.cpp
@@ -7,8 +7,11 @@ void sqlrconnection_svr::bindFormatCommand() {
 
 	dbgfile.debugPrint("connection",1,"bind format");
 
+	// get the bind format
 	const char	*bf=bindFormat();
-	uint16_t	bflen=(uint16_t)charstring::length(bf);
+
+	// send it to the client
+	uint16_t	bflen=charstring::length(bf);
 	clientsock->write(bflen);
 	clientsock->write(bf,bflen);
 	flushWriteBuffer();
