+++ src/connection/sqlrconnection/identify.cpp
@@ -7,8 +7,11 @@ void sqlrconnection_svr::identifyCommand() {
 
 	dbgfile.debugPrint("connection",1,"identify");
 
+	// get the identification
 	const char	*ident=identify();
-	uint16_t	idlen=(uint16_t)charstring::length(ident);
+
+	// send it to the client
+	uint16_t	idlen=charstring::length(ident);
 	clientsock->write(idlen);
 	clientsock->write(ident,idlen);
 	flushWriteBuffer();
