+++ src/connection/sqlrconnection/serverversion.cpp
@@ -7,8 +7,11 @@ void sqlrconnection_svr::serverVersionCommand() {
 
 	dbgfile.debugPrint("connection",1,"server version");
 
+	// get the server version
 	const char	*svrversion=SERVER_VERSION;
-	uint16_t	svrvlen=(uint16_t)charstring::length(svrversion);
+
+	// send it to the client
+	uint16_t	svrvlen=charstring::length(svrversion);
 	clientsock->write(svrvlen);
 	clientsock->write(svrversion,svrvlen);
 	flushWriteBuffer();
